Moved createSymbol
Former-commit-id: b081f0a4fe2a00eac60b7f16813160427d52e327
Former-commit-id: 09c1e49e67679c80ba10049a8e2201b0369ee81e

+++ include/sema/AntlrConverter.h
@@ -34,7 +34,7 @@ public:
     LogosMainFile* getMainFile(LogosParser::MainFileContext* ctx);
     LogosObjectFile* getObjFile(LogosParser::ObjectFileContext* ctx);
     LogosObject* getObject(LogosParser::ObjectFileContext* ctx);
-    LogosField *getField(LogosParser::ExplicitVarDecContext* varDec, const string& parentName, int position);
+    LogosField *getField(LogosParser::ExplicitVarDecContext* varDec, const string& parentName, size_t position);
     LogosUserFunc* getFunc(LogosParser::FuncImplementationContext* ctx);
     LogosStmt* getStmt(LogosParser::StatementContext* ctx);
     LogosFieldDef* getFieldDef(LogosParser::FieldDefContext* ctx);
